pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-harvestgym"
version = "0.1.0"
description = "HARvestGym: RL environment for training API-native web agents via HAR-guided exploration"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "requests>=2.31.0",
    "rank-bm25>=0.2.2",
    "openai>=1.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
]
embeddings = [
    "sentence-transformers>=3.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server", "server.tools"]

[tool.setuptools.package-data]
"*" = ["hars/*.har", "catalogs/*.json", "parameter_pools.json"]
